Added utils file for training, train_mlp_utils.hpp.
