+++ CMakeLists.txt
@@ -5,6 +5,7 @@ SET(ROOTDIR "${CMAKE_SOURCE_DIR}")
 INCLUDE_DIRECTORIES("${ROOTDIR}")
 
 add_definitions(-DMODEL_PATH="${CMAKE_SOURCE_DIR}/train/mlpack/")
+add_definitions(-DMLP_CONFIG_PATH="${CMAKE_SOURCE_DIR}/train/mlpack/model-regeneration/")
 
 find_package(PkgConfig REQUIRED)
 find_package(GTest REQUIRED)
