+++ train/mlpack/model-regeneration/mlp_config_utils.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <fstream>
 #include <iostream>
 #include <nlohmann/json.hpp>
